Voting and ToR discussion
I think we should update the ToR voting section to make it clear that the voting description there is for iTC-related votes. What I am intending here is that this would be for things impacting the iTC as a whole, and not specifically for things like voting on the documents. That isn't to say that this couldn't be used for those things, just that the expectation would be that the iTC may have other voting parameters for those instead.

So what I'm thinking is along the lines added here to make this more clear.