ToR Code of Conduct
So I was looking at some of the contributing guides and one in particular pointed to this site for a code of conduct:

https://www.contributor-covenant.org/version/2/0/code_of_conduct/

I thought this actually looked pretty good, and it can be modified as needed, they just ask for attribution, which is easy. I then looked at the ToR to see what was in there, and generally there is a match overall, but that is much more high level than the covenant.

What really got me though was the language about specific ISO things since the language is taken from the ISO docs. The iTCs aren't actually ISO organizations, so I don't think that should be there, even if the overall goal is something that would be acceptable under the ISO CC spec.

So here I have updated the ToR and also added the code of conduct. I'm somewhat concerned about the blocking sections, but I think that is really just accepted that if someone is not acting in an acceptable fashion that they would be kicked out.